chore(project): rename directories
- rename src/modules to src/core
