+++ src/core/auth/route-resolver.js
