+++ src/core/firebase/index.js
