+++ src/reducer.js
@@ -2,10 +2,10 @@ import { combineReducers } from 'redux';
 import { routeReducer } from 'react-router-redux';
 
 // Reducers
-import { authReducer } from 'modules/auth';
-import { firebaseReducer } from 'modules/firebase';
-import { notificationReducer } from 'modules/notification';
-import { tasksReducer } from 'modules/tasks';
+import { authReducer } from 'core/auth';
+import { firebaseReducer } from 'core/firebase';
+import { notificationReducer } from 'core/notification';
+import { tasksReducer } from 'core/tasks';
 
 
 export default combineReducers({
